Update short bio to make it more specific
Just updated my bio on ResearchGate, and reconsidered
that calling myself "spine surgeon" would be a bold statement,
so I decided to make this more specific. Also, I am starting
to realize that this dermatitis experience really was pivotal
in my career, so without stating it the picture in incomplete.

Also, still not sure if knowledge synthesis or evidence synthesis
works better for marketing.